Enable builders to have a hook for untagged sections
Previously, only tagged sections could have a `[tag]_begin` and `[tag]_end`,
but for some languages, such hook is desired also for untagged sections.
The `headline` hook could be used to open such sections,
but there was nothing to close them.

It is now possible to use `section_begin` and `section_end` in this case.
If these are defined, `headline` is no longer called,
and `section_begin` needs to be used instead.
If they are not defined, `headline` will work as usual,
so existing builders continue to work unchanged.